refactor and rename a bit.  still not totally happy with it
