+++ sdk/tests/test_utilities/TestCpuReferenceGraphExecutor.cpp
@@ -5,10 +5,10 @@
 #include <hipdnn_sdk/plugin/EnginePluginApi.h>
 #include <hipdnn_sdk/plugin/PluginApiDataTypes.h>
 #include <hipdnn_sdk/plugin/flatbuffer_utilities/GraphWrapper.hpp>
-#include <hipdnn_sdk/test_utilities/AdamTestStuff.hpp>
-#include <hipdnn_sdk/test_utilities/BatchnormSignatureRegistrar.hpp>
 #include <hipdnn_sdk/test_utilities/FlatbufferGraphTestUtils.hpp>
-#include <hipdnn_sdk/utilities/CpuReferenceGraphExecutor.hpp>
+#include <hipdnn_sdk/test_utilities/cpu_graph_executor/BatchnormBuilder.hpp>
+#include <hipdnn_sdk/test_utilities/cpu_graph_executor/BatchnormSignatureRegistrar.hpp>
+#include <hipdnn_sdk/test_utilities/cpu_graph_executor/CpuReferenceGraphExecutor.hpp>
 #include <hipdnn_sdk/utilities/Tensor.hpp>
 #include <hipdnn_sdk/utilities/TensorView.hpp>
 #include <hipdnn_sdk/utilities/UtilsBfp16.hpp>
@@ -18,75 +18,11 @@ using namespace hipdnn_sdk::test_utilities;
 using namespace hipdnn_sdk::data_objects;
 using namespace hipdnn_sdk::utilities;
 
-class TestAdamTestsStuff
+class TestCpuReferenceGraphExecutor
 {
 private:
 public:
 };
-
-TEST(TestAdamTestsStuff, BatchnormFwdInferenceFloat)
-{
-    Tensor<float> inputTensor({1, 3, 224, 224});
-    Tensor<float> outputTensor({1, 3, 224, 224});
-    Tensor<float> biasTensor({1, 3});
-    Tensor<float> scaleTensor({1, 3});
-    Tensor<float> meanTensor({1, 3});
-    Tensor<float> varianceTensor({1, 3});
-
-    BatchnormSignatureKey key{
-        .inputDataType = DataType::DataType_FLOAT,
-        .nodeAttributesType = NodeAttributes_BatchnormInferenceAttributes,
-    };
-
-    auto it = batchnormRegistry().find(key);
-    if(it != batchnormRegistry().end())
-    {
-        std::any inputAny = std::ref(inputTensor);
-        std::any scaleAny = std::ref(scaleTensor);
-        std::any biasAny = std::ref(biasTensor);
-        std::any meanAny = std::ref(meanTensor);
-        std::any varianceAny = std::ref(varianceTensor);
-        std::any outputAny = std::ref(outputTensor);
-
-        it->second(inputAny, scaleAny, biasAny, meanAny, varianceAny, outputAny, 1e-5);
-    }
-    else
-    {
-        throw std::runtime_error("No batchnorm implementation registered for this signature.");
-    }
-}
-
-TEST(TestAdamTestsStuff, BatchnormFwdInferenceHalf)
-{
-    Tensor<half> inputTensor({1, 3, 224, 224});
-    Tensor<half> outputTensor({1, 3, 224, 224});
-    Tensor<half> biasTensor({1, 3});
-    Tensor<half> scaleTensor({1, 3});
-    Tensor<half> meanTensor({1, 3});
-    Tensor<half> varianceTensor({1, 3});
-
-    BatchnormSignatureKey key{
-        .inputDataType = DataType::DataType_HALF,
-        .nodeAttributesType = NodeAttributes_BatchnormInferenceAttributes,
-    };
-
-    auto it = batchnormRegistry().find(key);
-    if(it != batchnormRegistry().end())
-    {
-        std::any inputAny = std::ref(inputTensor);
-        std::any scaleAny = std::ref(scaleTensor);
-        std::any biasAny = std::ref(biasTensor);
-        std::any meanAny = std::ref(meanTensor);
-        std::any varianceAny = std::ref(varianceTensor);
-        std::any outputAny = std::ref(outputTensor);
-
-        it->second(inputAny, scaleAny, biasAny, meanAny, varianceAny, outputAny, 1e-5);
-    }
-    else
-    {
-        throw std::runtime_error("No batchnorm implementation registered for this signature.");
-    }
-}
 
 template <typename T>
 hipdnnPluginDeviceBuffer_t
@@ -108,7 +44,7 @@ hipdnnPluginDeviceBuffer_t generateEmptyHostBuffer(TensorBase<T>& tensor, int ui
     return buffer;
 }
 
-TEST(TestAdamTestsStuff, Stuff2)
+TEST(TestCpuReferenceGraphExecutor, CanExecuteAGraphWithAFwdBatchnormNode)
 {
     unsigned int seed = std::random_device{}();
 
@@ -168,6 +104,6 @@ TEST(TestAdamTestsStuff, Stuff2)
         variantPack[deviceBuffer.uid] = deviceBuffer.ptr;
     }
 
-    hipdnn_sdk::utilities::CpuReferenceGraphExecutor::executeTheGraph(
+    hipdnn_sdk::test_utilities::CpuReferenceGraphExecutor::executeTheGraph(
         batchnormGraph, batchnormBuilder.GetSize(), variantPack);
 }
